requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
cachetools==3.1.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
flake8==3.7.7
future==0.17.1
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
kombu==4.5.0
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PySocks==1.6.8
python-telegram-bot==11.1.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.8
SQLAlchemy==1.3.1
urllib3==1.24.1
vine==1.3.0
